# src/AppBundle/CSPro/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Web API for synchronizing data and files from devices in the field to a web server.\n",
    "version": "2.0.0",
    "title": "CSPro Sync",
    "contact": {
      "name": "cspro@lists.census.gov"
    },
    "license": {
      "name": "Public domain"
    }
  },
  "host": "csweb.teleyah.com",
  "basePath": "/api/",
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json",
    "text/plain",
    "application/zip"
  ],
  "paths": {
...
    "/apps/{appname}": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "Get app package",
        "description": "Get application deployment package (csapb file) for named application",
        "operationId": "getApp",
        "produces": [
          "application/octet-stream"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appname",
            "description": "name of app to download",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "description": "Signature (MD5) of current application package specification on client. If same as signature of package on server 304 code is returned.",
            "type": "string"
          },
          {
            "in": "header",
            "name": "x-csw-package-build-time",
            "required": false,
            "description": "Build date/time of application package on client in RFC3339 format ( e.g. 1985-04-12T23:20:50.52Z). If server package is more recent then 304 code is returned.",
            "type": "string"
          },
          {
            "in": "header",
            "name": "x-csw-package-files",
            "required": false,
            "description": "List of application package files and signatures from JSON package spec on client gzcompressed and base 64 encoded. Files in this list with same signature as those in package on server are not sent in application package that is returned.",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "file"
            },
            "headers": {
              "Content-Length": {
                "description": "Size in bytes of downloaded file",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "304": {
            "description": "Client version of package same or newer than server version based on If-None-Match or x-csw-package-build-time header. No update needed."
          },
          "403": {
            "description": "Unable to read file"
          },
          "404": {
            "description": "File not found"
          }
        }
      },
      "put": {
        "tags": [
          "app"
        ],
        "summary": "Upload app package",
        "description": "Upload application package (csapb file) to server.",
        "operationId": "updateApp",
        "consumes": [
          "application/octect-stream"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appname",
            "description": "Name of app to upload",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "500": {
            "description": "Error writing file"
          }
        }
      },
      "delete": {
        "tags": [
          "app"
        ],
        "summary": "Delete app",
        "operationId": "deleteApp",
        "parameters": [
          {
            "in": "path",
            "name": "appname",
            "description": "The name of the app to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "App not found"
          }
        }
      }
    }
  },
...
}